Renamed package structure and add svg interface
