+++ src/main/Controller/SvgFileCreation.java
@@ -1,0 +1,8 @@
+package Controller;
+
+public interface SvgFileCreation {
+	public void createEmptyFile();
+	public void display();
+	public void modifyFile();
+	public void deleteFile();
+}
